Implement `--dry-run` flag. (#53)
* Implement dry-run.

* Simple test.

* Test the dryrun functionality in organiser.

* Move directory counting code into the test helpers package.

* Aaaand test the extraction with dry-run.

* Revert accidental deletion of some tests!

* Correct function name in output.


* Correct order of operations.


* Another wrong function name.


* Document config.


---------

+++ internal/extractor/extractor.go
@@ -49,6 +49,11 @@ func processTrack(f *zip.File, destination string) error {
 		return fmt.Errorf("error parsing music file name %s: %v", f.Name, err)
 	}
 
+	if o.Config.DryRun {
+		fmt.Printf("[Dry run] Extract track → %s\n", o.TrackDestination(track, destination))
+		return nil
+	}
+
 	// Actually extract and write the file
 	fd, err := os.Create(o.TrackDestination(track, destination))
 	if err != nil {
